requirements.txt
anthropic
notion-client
python-telegram-bot
chromadb
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
